feature(userInfo): oauth2로그인할 때 정보를 담기 위한 클래스 추가
